+++ src/startup/main_windows.cpp
@@ -2,44 +2,20 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "main_windows.h"
-
-#include <stdexcept>
-
 #include <windows.h>
 #include <winrt/base.h>
 
-#include <QApplication>
+#include <QCoreApplication>
 
 #include "log/log.h"
 #include "startup/windowsmessagehandler.h"
 #include "ui/darkthemeapplier_windows.h"
 #include "ui/systemcolorsprovider.h"
+#include "main_windows.h"
 #include "windowshelpers.h"
 
 namespace tremotesf {
-    namespace {
-        void onTerminate() {
-            const auto exception_ptr = std::current_exception();
-            if (exception_ptr) {
-                try {
-                    std::rethrow_exception(exception_ptr);
-                } catch (const std::exception& e) {
-                    warning().logWithException(e, "Unhandled exception");
-                } catch (const winrt::hresult_error& e) {
-                    warning().logWithException(e, "Unhandled exception");
-                } catch (...) {
-                    warning().log("Unhandled exception of unknown type");
-                }
-            }
-            std::abort();
-        }
-    }
-
-    WindowsLogger::WindowsLogger() {
-        initWindowsMessageHandler();
-        std::set_terminate(onTerminate);
-    }
+    WindowsLogger::WindowsLogger() { initWindowsMessageHandler(); }
 
     WindowsLogger::~WindowsLogger() { deinitWindowsMessageHandler(); }
 
@@ -59,7 +35,7 @@ namespace tremotesf {
         } catch (const std::system_error& e) {
             warning().log(e);
         }
-        const auto systemColorsProvider = SystemColorsProvider::createInstance(QApplication::instance());
+        const auto systemColorsProvider = SystemColorsProvider::createInstance(QCoreApplication::instance());
         applyDarkThemeToPalette(systemColorsProvider);
     }
 }